Fix Stanza Ack maximum and tests
The stanza ack code now uses a constexpr without numeric cast to
calculate the maximum, and test code uses the constant UINT32_MAX from
limits, both of which should result in the same number.

Test-Information:

Unit tests pass on macOS 10.14

+++ Swiften/Elements/StanzaAck.cpp
@@ -1,19 +1,17 @@
 /*
- * Copyright (c) 2013 Isode Limited.
+ * Copyright (c) 2013-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
 
 #include <Swiften/Elements/StanzaAck.h>
 
-#include <boost/numeric/conversion/cast.hpp>
-
 using namespace Swift;
 
 StanzaAck::~StanzaAck() {
 }
 
-void StanzaAck::setHandledStanzasCount(int i) {
-    handledStanzasCount = boost::numeric_cast<unsigned int>(i);
+void StanzaAck::setHandledStanzasCount(unsigned int i) {
+    handledStanzasCount = i;
     valid = true;
 }
